Update name and add tests
