+++ README.md
@@ -1,4 +1,4 @@
-# 3D IO Mesh Core
+# io3d 
 
 This repository provides a Rust library with Python bindings for converting elevation rasters to triangle meshes. The library exposes functions to build a mesh and export it to the PLY format for use in Blender or other tools.
