+++ pyproject.toml
@@ -3,11 +3,12 @@ requires = ["maturin>=1.2,<2.0"]
 build-backend = "maturin"
 
 [project]
-name = "rust_core_mesh"
+name = "io3d"
 version = "0.1.0"
 requires-python = ">=3.8"
 dependencies = [
     "maturin>=1.8.6",
     "numpy>=1.24.4",
+    "pytest>=8.3.5",
     "rasterio>=1.3.11",
 ]
