+++ src/lib.rs
@@ -22,7 +22,7 @@ fn export_ply(mesh: &Mesh, path: &str) -> PyResult<()> {
 }
 
 #[pymodule]
-fn rust_core_mesh(_py: Python, m: &PyModule) -> PyResult<()> {
+fn io3d(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<Vertex>()?;
     m.add_class::<Face>()?;
     m.add_class::<Mesh>()?;
